Add a ignore path for texture killing

+++ Interface/AddOns/NDui/Core/Functions.lua
@@ -381,7 +381,7 @@ do
 		if self.GetNumRegions then
 			for i = 1, self:GetNumRegions() do
 				local region = select(i, self:GetRegions())
-				if region and region.IsObjectType and region:IsObjectType("Texture") then
+				if region and region.IsObjectType and region:IsObjectType("Texture") and not region.isIgnored then
 					if kill and type(kill) == "boolean" then
 						B.HideObject(region)
 					elseif tonumber(kill) then
